Move all platforms into a seperate file
